more MSWin32 workarounds
So this is bad.  I admit it!  However, tbh my use case for Alt is usually
just for it to work for long enough to get some cpantesters results.
If EUMM changes and breaks our workaround then we'll just have to
revisit this later maybe.  Also the workaround only needs and only
is applied for windows, and life is mostly terrible there anyway
for Perl developers.